+++ field.c
@@ -1107,7 +1107,7 @@ set_element(long num, char *s, long len, NODE *n)
 
 	it = make_string(s, len);
 	it->flags |= USER_INPUT;
-	sub = make_number((double) (num));
+	sub = make_number(num);
 	assoc_set(n, sub, it);
 }
 
@@ -1178,7 +1178,7 @@ do_split(int nargs)
 		 */
 		tmp = POP_SCALAR();
 		DEREF(tmp);
-		return make_number((double) 0);
+		return make_number(0.0);
 	}
 
 	if ((sep->flags & REGEX) != 0)
@@ -1329,7 +1329,7 @@ do_patsplit(int nargs)
 		/*
 		 * Skip the work if first arg is the null string.
 		 */
-		tmp = make_number((double) 0);
+		tmp = make_number(0.0);
 	} else {
 		rp = re_update(sep);
 		s = src->stptr;
